ad validation

+++ src/main/java/by/gradomski/apartments/command/impl/NewAdCommand.java
@@ -8,6 +8,7 @@ import by.gradomski.apartments.entity.User;
 import by.gradomski.apartments.exception.ServiceException;
 import by.gradomski.apartments.service.impl.AdServiceImpl;
 import by.gradomski.apartments.service.impl.ApartmentServiceImpl;
+import by.gradomski.apartments.validator.AdvertisementValidator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -15,6 +16,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 import static by.gradomski.apartments.command.PagePath.*;
 
@@ -25,9 +28,10 @@ public class NewAdCommand implements Command {
     private static final String TITLE = "title";
     private static final String PRICE = "price";
     private static final String ADVERTISEMENT_LIST = "advertisementList";
+    private static final String FALSE = "false";
 
     @Override
-    public String execute(HttpServletRequest request) {
+    public String execute(HttpServletRequest request) {         //TODO(through transaction)
         log.debug("start execute method");
         String page;
         HttpSession session = request.getSession(false);
@@ -41,33 +45,49 @@ public class NewAdCommand implements Command {
             long apartmentId = Long.parseLong((String) session.getAttribute(APARTMENT_ID));
             AdServiceImpl adService = AdServiceImpl.getInstance();
             ApartmentServiceImpl apartmentService = ApartmentServiceImpl.getInstance();
+            Map<String, String> validationResult = AdvertisementValidator.isValid(title, price);
             try {
-                long newAdvertisementId = adService.addAdvertisement(title, adAuthor, price, apartmentId);
-                if(newAdvertisementId > 0){
-                    boolean updateStatusResult = apartmentService.
-                            updateApartmentStatus(apartmentId, ApartmentStatus.IN_DEMAND);
-                    if(updateStatusResult){
-                        List<Apartment> updatedApartmentList = apartmentService.getApartmentsByOwner(adAuthor.getId());
-                        session.removeAttribute(APARTMENT_ID);
-                        session.setAttribute("apartmentList", updatedApartmentList); //TODO(change to request.attr?)
-                        Ad newAd = AdServiceImpl.getInstance().getAdById(newAdvertisementId);
-                        Object obj = request.getServletContext().getAttribute(ADVERTISEMENT_LIST);
-                        log.debug("obj from servletContext ADVERTISEMENT_LIST: " + obj);    //FIXME(if this first ad -> obj == null?)
-                        List<Ad> adList = (List<Ad>) obj;
-                        adList.add(newAd);
-                        request.getServletContext().setAttribute(ADVERTISEMENT_LIST, adList);
-                        page = ESTATE;
-                    } else {
-                        log.info("Advertisement was added; can't change apartment status");
-                        boolean deleteResult = adService.deleteAd(newAdvertisementId);
-                        if(!deleteResult){
-                            log.info("can't delete ad: id=" + newAdvertisementId);
+                if(!validationResult.containsValue(FALSE)){
+                    long newAdvertisementId = adService.addAdvertisement(title, adAuthor, price, apartmentId);
+                    if (newAdvertisementId > 0) {
+                        boolean updateStatusResult = apartmentService.
+                                updateApartmentStatus(apartmentId, ApartmentStatus.IN_DEMAND);
+                        if (updateStatusResult) {
+                            List<Apartment> updatedApartmentList = apartmentService.getApartmentsByOwner(adAuthor.getId());
+                            session.removeAttribute(APARTMENT_ID);
+                            session.setAttribute("apartmentList", updatedApartmentList); //TODO(change to request.attr?)
+                            Ad newAd = AdServiceImpl.getInstance().getAdById(newAdvertisementId);
+                            Object obj = request.getServletContext().getAttribute(ADVERTISEMENT_LIST);
+                            log.debug("obj from servletContext ADVERTISEMENT_LIST: " + obj);    //FIXME(if this first ad -> obj == null?)
+                            List<Ad> adList = (List<Ad>) obj;
+                            adList.add(newAd);
+                            request.getServletContext().setAttribute(ADVERTISEMENT_LIST, adList);
+                            page = ESTATE;
+                        } else {
+                            log.info("Advertisement was added; can't change apartment status");
+                            boolean deleteResult = adService.deleteAd(newAdvertisementId);
+                            if (!deleteResult) {
+                                log.info("can't delete ad: id=" + newAdvertisementId);
+                            }
+                            request.setAttribute("newAdErrorMessage", "Some errors occurred. Try again");
+                            page = NEW_AD;
                         }
-                        request.setAttribute("newAdErrorMessage", "Some errors occurred. Try again");
+                    } else {
+                        request.setAttribute("newAdErrorMessage", "Title and price are required fields. Try again");
                         page = NEW_AD;
                     }
                 } else {
-                    request.setAttribute("newAdErrorMessage", "Title and price are required fields. Try again");
+                    String failReason = defineFalseKey(validationResult);
+                    switch (failReason) {
+                        case TITLE:
+                            log.debug("incorrect title: " + title);
+                            request.setAttribute("titleError", true);
+                            break;
+                        case PRICE:
+                            log.debug("incorrect price: " + price);
+                            request.setAttribute("priceError", true);
+                            break;
+                    }
                     page = NEW_AD;
                 }
             } catch (ServiceException e) {
@@ -76,5 +96,15 @@ public class NewAdCommand implements Command {
             }
         }
         return page;
+    }
+
+    private String defineFalseKey(Map<String, String> map){
+        Optional<String> optionalResult = map.entrySet()
+                .stream()
+                .filter(entry -> FALSE.equals(entry.getValue()))
+                .map(Map.Entry::getKey)
+                .findFirst();
+
+        return optionalResult.get();
     }
 }
